Fixed incorrect int value
This commit fixes the open issue, where this error message is generated
in the apache error.log when data is written to the database.

[Sun Nov 08 07:54:47.014785 2015] [cgi:error] [pid 33171] [client
192.168.20.196:40033] AH01215:
/usr/local/lib/python2.7/dist-packages/pymysql/cursors.py:146: Warning:
Incorrect integer value: '' for column 'id' at row 1, referer:
http://192.168.20.196/index.py?module=enterdb

The issue was I was passing an empty string, which worked, but
generated the warning.  I am now passing a 0 (zero), which works as
well and does not generate the warning.