0.4.1
Added functionality to display the local time in the server information
box.  This will show the current time of the system which can be used
when comparing to the timestamps for data saved to the MySQL database.
This was added because of demonstrating backup and restore processes
for application aware backup/restore products.  I also modified how the
timestamps get time to make it the same as the server information box.