Menu Updates - Fixed Name to match README
I noticed I needed to rename the filenames for the commands to call the
menu back up.  It is under the assumption that the README was followed
and the user copied the correct menu_***.sh depending on what server
role the machine is functioning.  Changing this in the code will not
require not the user edit the bash scripts to get the menus working.