0.4.3 - Display Client Information
This commits added functionality requested (Issue #10) to have
information regarding the client making the HTTP request displayed.
The idea is to be able to see when load balancers are in play and also
when they are modifying the traffic to a site.

Displaying the X_Forwarded_For information has not been tested with a
load balancing solution at this time.  I am not 100% certain this will
work, but if not, I can work on it if I can get my load-balancer system
fixed in the near future.

Enjoy!