0.5.0
I added error checking functionality to the site.  If the application
tier is unavailable, an error message stating so will be displayed.  If
the MySQL server database is unavailable, then a different message is
displayed indicating that.  This will help with demonstrations when you
are breaking communication to the different tiers.

Also removed the logic when the page was cleared, the app server
information was not displayed.  I have that functionality in the app
now even though you don't "really" connect to the app server.  This
will be useful will you break communication to the app and db tiers.