Auto Installer 0.1.0
This is the start of a auto installation script that will eventually be
used to ease the deployment of this app.

I’ve also included a script written by Luis Chanu for this app to help
with easing the pain of making Apache config modifications.  This
script, a2appdemo.sh, I have not tested yet.