Install Script 0.1.1
I have made some changes to the install script and beefed up the menu
some.  I ended up doing some nested case statements because the way the
menu was formatting if there were too many characters in the menu
option.  This is a little of a hack, but it works.